fix: resolve build error by passing searchParams as props, and separating client components
