+++ src/app/auth/callback/styles.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/utils.module.scss';
+@import "../../../styles/utils.module.scss";
 
 .container {
   @extend .container;
@@ -7,28 +7,4 @@
   @extend .alignItemsCenter;
   min-height: 100vh;
   width: 100%;
-}
-
-.content {
-  @extend .flexCol;
-  @extend .alignItemsCenter;
-  padding: 3rem;
-  text-align: center;
-  background: white;
-}
-
-.title {
-  @extend .heading1;
-  margin-bottom: 2rem;
-  color: #24292e;
-}
-
-.spinner {
-  @extend .animate-spin;
-  border-radius: 50%;
-  width: 3rem;
-  height: 3rem;
-  border: 4px solid #f6f8fa;
-  border-top: 4px solid #24292e;
-  margin: 0 auto;
 }
