+++ src/components/Auth/CallbackClient/styles.module.scss
@@ -1,0 +1,25 @@
+@import "../../../styles/utils.module.scss"
+
+.content {
+  @extend .flexCol;
+  @extend .alignItemsCenter;
+  padding: 3rem;
+  text-align: center;
+  background: white;
+}
+
+.title {
+  @extend .heading1;
+  margin-bottom: 2rem;
+  color: #24292e;
+}
+
+.spinner {
+  @extend .animate-spin;
+  border-radius: 50%;
+  width: 3rem;
+  height: 3rem;
+  border: 4px solid #f6f8fa;
+  border-top: 4px solid #24292e;
+  margin: 0 auto;
+}
